Working Map
I feel really good about this, I found the previous error that was stopping the Emulator from working. Basically, I had accidentally deleted an ID in the XML file that was called by another file. Now that's fixed. One Last Thing: We need a button to get the Map up and running. But everything is building/running in the emulator :grinning: